Fixed some model issues

+++ src/main/java/org/ukma/spring/crooodle/model/Booking.java
@@ -27,6 +27,6 @@ public class Booking {
 
     private LocalDate startDate;
     private LocalDate endDate;
-    private double totalPrice;
+    private int totalPrice;
     private String status;
 }
